Having used the NAG Fortran compiler on a small part of DELWAQ (the processes library) I came across a number of issues that I have repaired using the compiler report: - Unused variables (removed) - Superfluous SAVE attribute - A missing comma (actually an error!), leading to an odd variable name - Very long lines (> 132) - REAL*4/REAL*8 converted - Subexpressions like a**-1

+++ src/engines_gpl/waq/packages/waq_kernel/src/waq_process/mac3du.f90
@@ -55,7 +55,7 @@
       real(4) totaldepth  ! i  total depth water column                           (m)
       real(4) nh4s12      ! i  concentration of NH4 in S12 bottom (pores)         (g/m3)
       real(4) locseddept  ! i  sediment layer depth to bottom of segment          (m)
-      real(4) ibotseg     ! i  bottom segment number                              (-)
+      integer ibotseg     ! i  bottom segment number                              (-)
       real(4) FrBmLay     ! i  actual height sm                                   (m)
       real(4) rootdesm01  ! i  rooting depth sm01                                 (m)
       real(4) poros       ! i  volumetric porosity                                (-)
